fixed redundant code, and more

+++ src/main/java/com/hamusuke/twitter4mc/gui/screen/ErrorScreen.java
@@ -39,7 +39,7 @@ public class ErrorScreen extends ParentalScreen {
         this.drawCenteredString(this.font, this.title.asFormattedString(), this.width / 2, 20, Formatting.WHITE.getColorValue());
         List<String> list = this.font.wrapStringToWidthAsList(this.errormsg, this.width / 3);
         for (int i = 0; i < list.size(); i++) {
-            this.font.drawWithShadow(list.get(i), this.width / 3, 50 + i * 10, Formatting.WHITE.getColorValue());
+            this.font.drawWithShadow(list.get(i), (float) this.width / 3, 50 + i * 10, Formatting.WHITE.getColorValue());
         }
     }
 }
